requirements.txt
fastapi==0.119.0
uvicorn==0.38.0
requests==2.32.5
python-dotenv==1.1.1
sentence-transformers==5.1.1
torch==2.9.0
faiss-cpu==1.12.0
transformers==4.57.1
numpy==2.3.4